Pull to resync.

+++ win/share/tiletext.c
@@ -1,4 +1,4 @@
-/* NetHack 3.5	tiletext.c	$NHDT-Date: 1429484196 2015/04/19 22:56:36 $  $NHDT-Branch: master $:$NHDT-Revision: 1.6 $ */
+/* NetHack 3.5	tiletext.c	$NHDT-Date: 1430640200 2015/05/03 08:03:20 $  $NHDT-Branch: master $:$NHDT-Revision: 1.7 $ */
 /* NetHack 3.5	tiletext.c	$Date: 2009/05/06 10:59:03 $  $Revision: 1.4 $ */
 /*	SCCS Id: @(#)tiletext.c	3.5	1999/10/24	*/
 /* NetHack may be freely redistributed.  See license for details. */
@@ -286,8 +286,8 @@ const char *type;
 		/* Fill placeholder with noise */
 		if ( !placeholder_init ) {
 		    placeholder_init++;
-		    for ( i=0; i<sizeof(placeholder); i++ )
-			((char*)placeholder)[i]=i%256;
+		    for (i = 0; i < (int)sizeof placeholder; i++)
+			((char *)placeholder)[i] = i % 256;
 		}
 
 		read_text_colormap(tile_file);
